# services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Ocserv User management Api server.",
        "title": "Ocserv User management Example Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/customers/setup/ios": {
            "post": {
                "description": "Create Cisco Secure Client iOS certificate import and connection creation URIs using ocserv username/password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Create customer iOS setup links",
                "parameters": [
                    {
                        "description": "customer username and password (same ocserv account).",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/customer.SummaryData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/customer.IOSSetupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
        "/customers/setup/ios/certificate/{token}": {
            "get": {
                "description": "Download customer's PKCS#12 certificate using a short-lived iOS setup token",
                "produces": [
                    "application/x-pkcs12"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Download customer iOS setup certificate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iOS setup certificate token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user.p12",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "customer.IOSSetupResponse": {
            "type": "object",
            "required": [
                "certificate_import_uri",
                "certificate_password",
                "connection_create_uri",
                "connection_name",
                "expires_at",
                "server_address",
                "server_port"
            ],
            "properties": {
                "certificate_import_uri": {
                    "type": "string"
                },
                "certificate_password": {
                    "type": "string"
                },
                "connection_create_uri": {
                    "type": "string"
                },
                "connection_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "server_address": {
                    "type": "string"
                },
                "server_port": {
                    "type": "integer"
                }
            }
        },
...
        "customer.SummaryData": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                }
            }
        },
...
        "middlewares.TooManyRequests": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "required": [
                "error",
                "message"
            ],
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}